requirements.txt
cloudinary
Flask
opencv-python
Pillow
requests
python-dotenv